app: Normalize kind property on retrieved items
One of the warts in the kubernetes REST API is that depending on how
you get an item, it may have a kind property or not. Clients are
expected to infer the kind from type they requested.

The topology view processes items generically, and this the kind
property plays a large role in that.